requirements.txt
# 核心Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket支持
websockets==12.0

# Redis客户端
redis==5.0.1

# 数据验证
pydantic==1.10.13
email-validator==2.1.0

# 安全认证
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# 工具库
python-dotenv==1.0.0
python-multipart==0.0.6
schedule==1.2.0
